+++ lib/readline.js
@@ -44,7 +44,7 @@ const {
   isFullWidthCodePoint,
   kUTF16SurrogateThreshold,
   stripVTControlCharacters
-} = require('internal/readline');
+} = require('internal/readline/utils');
 
 const { clearTimeout, setTimeout } = require('timers');
 const {
