+++ node.gyp
@@ -169,7 +169,7 @@
       'lib/internal/process/report.js',
       'lib/internal/process/task_queues.js',
       'lib/internal/querystring.js',
-      'lib/internal/readline.js',
+      'lib/internal/readline/utils.js',
       'lib/internal/repl.js',
       'lib/internal/repl/await.js',
       'lib/internal/repl/history.js',
